[mdal] move GUI related code to a new file
